Move table components into tables path for organizational clarity
