+++ app/components/dashboard/show/works_list_component.html.erb
@@ -1,5 +1,5 @@
 <h3 class="h3 mb-4"><%= label %></h3>
-<%= render Elements::TableComponent.new(
+<%= render Elements::Tables::TableComponent.new(
       id: 'worksTable',
       classes: 'table table-striped table-sm table-work mb-5',
       header_classes: 'table-light',
