+++ app/components/elements/tables/table_body_component.html.erb
