+++ app/components/elements/tables/table_component.html.erb
@@ -1,0 +1,7 @@
+<%= tag.table class: classes, id: do %>
+  <% if label.present? %>
+    <caption><%= label %></caption>
+  <% end %>
+  <%= render Elements::Tables::TableHeaderComponent.new(classes: header_classes, headers: headers) if headers %>
+  <%= render Elements::Tables::TableBodyComponent.new(rows: rows) if rows %>
+<% end %>
