+++ spec/components/elements/tables/table_row_component_spec.rb
@@ -2,7 +2,7 @@
 
 require 'rails_helper'
 
-RSpec.describe Elements::TableRowComponent, type: :component do
+RSpec.describe Elements::Tables::TableRowComponent, type: :component do
   it 'renders the row' do
     render_inline(described_class.new(values: ['First value', 'Second value']))
     expect(page).to have_css('td', count: 2)
